Blog/requirements.txt
Package             Version
------------------- -------
asgiref             3.2.10
Django              3.1.2
djangorestframework 3.12.1
pip                 20.1.1
psycopg2            2.8.6
pytz                2020.1
setuptools          47.1.0
sqlparse            0.4.1
